pkgs/bay/pyproject.toml
[project]
name = "bay"
version = "0.1.0"
description = "Bay - Orchestration layer for Ship containers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Web framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    
    # Database
    "sqlmodel>=0.0.14",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",
    
    # Docker
    "aiodocker>=0.21.0",
    
    # HTTP client
    "httpx>=0.26.0",
    
    # Config & validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.0",
    
    # Auth (Phase 1 简化)
    "python-jose[cryptography]>=3.3.0",
    
    # Utils
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.13"
strict = true
